+++ src/Modules/Dependencies.php
@@ -30,6 +30,7 @@ class Dependencies extends \RubbishThorClone
         $this->command('update', 'Updates dependencies to their latest versions. Use deps update [type]/[name] to update a specific dependency', $inspections_host_option);
         $this->command('migrate', 'Converts legacy plugins file to whippet.json');
         $this->command('validate', 'Validate whippet.json and whippet.lock files');
+        $this->command('describe', 'List dependencies and their versions');
     }
 
     private function exitIfError(\Result\Result $result)
@@ -82,6 +83,13 @@ class Dependencies extends \RubbishThorClone
         $dir = $this->getDirectory();
         $validator = new \Dxw\Whippet\Dependencies\Validator($this->factory, $dir);
         $this->exitIfError(($validator->validate()));
+    }
+
+    public function describe()
+    {
+        $dir = $this->getDirectory();
+        $describer = new \Dxw\Whippet\Dependencies\Describer($this->factory, $dir);
+        $this->exitIfError(($describer->describe()));
     }
 
     private function inspectionChecker()